etl clinical: move custom errors to shared module
Refactor the custom "Unknown*" errors into a shared module for reuse by
other etl commands.
